app/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
boto3==1.28.72
botocore==1.31.72
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
diagrams==0.23.3
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.2
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.0
graphviz==0.20.1
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
orjson==3.9.9
passlib==1.7.4
prompt-toolkit==3.0.39
psycopg2==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tqdm==4.66.1
typed-ast==1.5.5
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
yarl==1.9.2
